requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
tensorboard
nibabel
os
glob

